refactoring: types & cleaning

+++ packages/syngrisi/src/server/controllers/runs.controller.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import httpStatus from 'http-status';
 import { EJSON } from 'bson';
@@ -21,6 +20,8 @@ const get = catchAsync(async (req: ExtRequest, res: Response) => {
 
 const remove = catchAsync(async (req: ExtRequest, res: Response) => {
     const { id } = req.params;
+    if (!req.user) throw new Error("req.user is empty");
+
     const result = await runService.remove(id, req?.user);
     res.send(result);
 });
